Use the new sass API and migrate to sass-embedded
