+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@wide/forge",
-  "version": "2.3.3",
+  "version": "3.0.0-alpha.1",
   "description": "Zero-based configuration builder for frontend integration projects.",
   "license": "MIT",
   "author": "Aymeric Assier (https://github.com/myeti)",
@@ -17,7 +17,7 @@
     "forge": "./lib/cli.js"
   },
   "engines": {
-    "node": ">=12.18.0"
+    "node": ">=14.19.0"
   },
   "scripts": {
     "build": "babel src/ -d lib/ --presets=@babel/preset-env --plugins=@babel/plugin-syntax-object-rest-spread,@babel/plugin-transform-runtime",
@@ -40,7 +40,7 @@
     "parcel-bundler": "^1.12.5",
     "postcss": "^8.3.11",
     "rimraf": "^3.0.2",
-    "sass": "^1.43.4",
+    "sass-embedded": "^1.50.0",
     "svg-sprite": "^1.5.3",
     "twig": "^1.15.4",
     "yargs": "^17.2.1"
